+++ src/form/_example/clear-validate.vue
@@ -13,7 +13,7 @@
     </t-form-item>
 
     <t-form-item label="邮箱" name="email">
-      <t-input v-model="formData.email"></t-input>
+      <t-auto-complete v-model="formData.email" :options="emailOptions" filterable></t-auto-complete>
     </t-form-item>
 
     <t-form-item label="性别" name="gender">
@@ -58,7 +58,7 @@
   </t-form>
 </template>
 <script setup>
-import { ref, reactive } from 'vue';
+import { ref, reactive, computed } from 'vue';
 import { MessagePlugin } from 'tdesign-vue-next';
 
 const form = ref(null);
@@ -75,6 +75,13 @@ const formData = reactive({
   },
   course: [],
 });
+const emailSuffix = ['@qq.com', '@163.com', '@gmail.com'];
+const emailOptions = computed(() => {
+  const emailPrefix = formData.email.split('@')[0];
+  if (!emailPrefix) return [];
+
+  return emailSuffix.map((suffix) => emailPrefix + suffix);
+});
 
 const courseOptions = [
   { label: '语文', value: '1' },
@@ -82,6 +89,7 @@ const courseOptions = [
   { label: '英语', value: '3' },
   { label: '体育', value: '4' },
 ];
+
 const options = [
   { label: '计算机学院', value: '1' },
   { label: '软件学院', value: '2' },
@@ -123,7 +131,6 @@ const rules = {
     },
   ],
 };
-
 const onReset = () => {
   MessagePlugin.success('重置成功');
   console.log('formData', formData.value);