+++ src/grid/type.ts
@@ -11,7 +11,7 @@ export interface TdRowProps {
    */
   align?: 'start' | 'end' | 'center' | 'stretch' | 'baseline' | 'top' | 'middle' | 'bottom';
   /**
-   * 栅格间隔，示例：`{ xs: 8, sm: 16, md: 24}`。当数据类型为 Number 和 Object 时，用于指定横向间隔。当数据类型为数组时，第一个参数为横向间隔，第二个参数为纵向间隔， [水平间隔, 垂直间隔]
+   * 栅格间隔，示例：`{ xs: 8, sm: 16, md: 24, lg: 32, xl: 32, xxl: 40 }`。当数据类型为 Number 和 Object 时，用于指定横向间隔。当数据类型为数组时，第一个参数为横向间隔，第二个参数为纵向间隔， [水平间隔, 垂直间隔]
    * @default 0
    */
   gutter?: number | GutterObject | Array<GutterObject | number>;
@@ -66,7 +66,6 @@ export interface TdColProps {
   sm?: number | BaseColProps;
   /**
    * 栅格占位格数，为 0 时相当于 display: none
-   * @default 12
    */
   span?: number;
   /**
@@ -92,6 +91,9 @@ export interface GutterObject {
   xs: number;
   sm: number;
   md: number;
+  lg: number;
+  xl: number;
+  xxl: number;
 }
 
 export interface BaseColProps {