+++ src/menu/submenu.tsx
@@ -11,6 +11,7 @@ import {
   toRefs,
   reactive,
   nextTick,
+  Transition,
 } from 'vue';
 import props from './submenu-props';
 import { renderContent, renderTNodeJSX } from '../utils/render-tnode';
@@ -21,6 +22,7 @@ import { usePrefixClass } from '../hooks/useConfig';
 import { Popup, PopupPlacement } from '../popup';
 import isFunction from 'lodash/isFunction';
 import { TdSubmenuProps } from './type';
+import useCollapseAnimation from '../hooks/useCollapseAnimation';
 
 export default defineComponent({
   name: 'TSubmenu',
@@ -50,6 +52,7 @@ export default defineComponent({
     const popupWrapperRef = ref<HTMLElement>();
     const subPopupRef = ref<HTMLElement>();
     const submenuRef = ref<HTMLElement>();
+    const transitionClass = usePrefixClass('slide-down');
     useRipple(submenuRef, rippleColor);
 
     const classes = computed(() => [
@@ -215,6 +218,8 @@ export default defineComponent({
       isNested,
       classes,
       subClass,
+      isOpen,
+      transitionClass,
       arrowClass,
       overlayInnerClassName,
       overlayClassName,
@@ -308,6 +313,8 @@ export default defineComponent({
         parent = parent.parent;
       }
 
+      const { beforeEnter, enter, afterEnter, beforeLeave, leave, afterLeave } = useCollapseAnimation();
+
       const needRotate = this.mode === 'popup' && this.isNested;
 
       const normalSubmenu = [
@@ -321,9 +328,19 @@ export default defineComponent({
             />
           )}
         </div>,
-        <ul class={this.subClass} style={{ '--padding-left': `${paddingLeft}px` }}>
-          {child}
-        </ul>,
+        <Transition
+          name={this.transitionClass}
+          onBeforeEnter={beforeEnter}
+          onEnter={enter}
+          onAfterEnter={afterEnter}
+          onBeforeLeave={beforeLeave}
+          onLeave={leave}
+          onAfterLeave={afterLeave}
+        >
+          <ul v-show={this.isOpen} class={this.subClass} style={{ '--padding-left': `${paddingLeft}px` }}>
+            {child}
+          </ul>
+        </Transition>,
       ];
 
       const triggerElement = [