+++ src/table/_example/custom-col-button.vue
@@ -18,6 +18,7 @@
     <!-- 2. displayColumns 动态设置显示哪些列，受控属性，支持 displayColumns.sync 语法糖 -->
     <!-- 3. onDisplayColumnsChange 当前显示列发生变化时触发 -->
     <!-- 4. 如果希望顶部内容 和 列配置按钮 保持在同一行，可将内容放在 topContent，并调整按钮父元素宽度(CSS) -->
+    <!-- 5. resizable and tableLayout: fixed is suggested  -->
     <!-- 受控用法，示例代码有效，勿删  -->
     <t-table
       v-model:displayColumns="displayColumns"
@@ -32,8 +33,8 @@
       }"
       :pagination="{ defaultPageSize: 5, defaultCurrent: 1, total: 100 }"
       :bordered="bordered"
-      table-layout="auto"
       stripe
+      resizable
       @column-change="onColumnChange"
     >
     </t-table>